requirement.txt
streamlit
opencv-python
numpy
matplotlib
requests
Pillow
transformers
opencv-python-headless
